Decouple personal information from CV
