+++ vitae/features/researchers/model/researcher.py
@@ -8,6 +8,7 @@ from ._shared import optional
 from .address import Address
 from .cv import Curriculum
 from .external import Lattes, Orcid
+from .personal import Person
 
 
 @attrs.frozen
@@ -25,13 +26,15 @@ class ExternalLink:
 
 @attrs.frozen
 class Researcher:
+    this: Person
+    curriculum: Curriculum
     links: ExternalLink
     address: Address | None
-    curriculum: Curriculum
 
     @classmethod
     def from_table(cls, table: tables.Researcher):
         return cls(
+            this=Person.from_table(table),
             links=ExternalLink.from_table(table),
             address=optional(table.address, Address.from_table),
             curriculum=Curriculum.from_table(table),
